+++ apps/InfoStation/setup.js
@@ -4,11 +4,11 @@ import * as constCommon from '../js/constellation_app_common.js'
 import * as constFileSelect from '../js/constellation_file_select_modal.js'
 import * as constSetup from '../js/constellation_setup_common.js'
 
-function clearDefinitionInput (full = true) {
-  // Clear all input related to a defnition
+function initializeDefinition () {
+  // Create a blank definition at save it to workingDefinition.
 
-  if (full === true) {
-  // Get a new temporary uuid
+  return new Promise(function (resolve, reject) {
+    // Get a new temporary uuid
     constCommon.makeHelperRequest({
       method: 'GET',
       endpoint: '/uuid/new'
@@ -41,7 +41,16 @@ function clearDefinitionInput (full = true) {
           }
         })
         constSetup.previewDefinition()
+        resolve()
       })
+  })
+}
+
+async function clearDefinitionInput (full = true) {
+  // Clear all input related to a defnition
+
+  if (full === true) {
+    await initializeDefinition()
   }
 
   // Language add
@@ -70,9 +79,8 @@ function clearDefinitionInput (full = true) {
     gradient_color_2: '#719abf'
   })
 
-  Array.from(document.querySelectorAll('.font-select')).forEach((el) => {
-    el.value = el.getAttribute('data-default')
-  })
+  constSetup.resetAdvancedFontPickers()
+
   Array.from(document.querySelectorAll('.text-size-slider')).forEach((el) => {
     el.value = 0
   })
@@ -120,10 +128,13 @@ function editDefinition (uuid = '') {
     document.querySelector('#colorPicker_' + key).dispatchEvent(new Event('input', { bubbles: true }))
   })
 
-  // Set the appropriate values for the font selects
-  Object.keys(def.style.font).forEach((key) => {
-    $('#fontSelect_' + key).val(def.style.font[key])
-  })
+  // Set the appropriate values for the advanced font pickers
+  if ('font' in def.style) {
+    Object.keys(def.style.font).forEach((key) => {
+      const picker = document.querySelector(`.AFP-select[data-path="style>font>${key}"`)
+      constSetup.setAdvancedFontPicker(picker, def.style.font[key])
+    })
+  }
 
   // Set the appropriate values for the text size sliders
   Object.keys(def.style.text_size).forEach((key) => {
@@ -558,39 +569,6 @@ function deleteInfoStationTab (lang, uuid) {
   $('.infostation-tab').click()
 }
 
-function onFontUploadChange () {
-  // Classed when the user selects font files to upload
-
-  const fileInput = $('#uploadFontInput')[0]
-  const files = fileInput.files
-  const formData = new FormData()
-
-  $('#uploadFontName').html('Uploading')
-
-  Object.keys(files).forEach((key) => {
-    const file = files[key]
-    formData.append('files', file)
-  })
-
-  const xhr = new XMLHttpRequest()
-  xhr.open('POST', '/uploadContent', true)
-
-  xhr.onreadystatechange = function () {
-    if (this.readyState !== 4) return
-    if (this.status === 200) {
-      const response = JSON.parse(this.responseText)
-
-      if ('success' in response) {
-        $('#uploadFontName').html('Upload new')
-        populateFontSelects()
-      }
-    } else if (this.status === 422) {
-      console.log(JSON.parse(this.responseText))
-    }
-  }
-  xhr.send(formData)
-}
-
 function onFlagUploadChange (lang) {
   // Called when the user selects a flag image file to upload
 
@@ -665,61 +643,6 @@ function onAttractorFileChange () {
   constSetup.previewDefinition(true)
 }
 
-function populateFontSelects () {
-  // Get a list of all the content and add the available font files to the appropriate selects.
-
-  const types = ['header', 'body', 'section-header', 'button']
-  $('.font-select').empty()
-
-  // First, search the content directory for any user-provided fonts
-  constCommon.makeHelperRequest({
-    method: 'GET',
-    endpoint: '/getAvailableContent'
-  })
-    .then((result) => {
-      types.forEach((type) => {
-        // First, add the default
-        const defaultFont = document.createElement('option')
-        defaultFont.value = $('#fontSelect_' + type).data('default')
-        defaultFont.innerHTML = 'Default'
-        $('#fontSelect_' + type).append(defaultFont)
-
-        const header = document.createElement('option')
-        header.value = 'User-provided'
-        header.innerHTML = 'User-provided'
-        header.setAttribute('disabled', true)
-        $('#fontSelect_' + type).append(header)
-
-        result.all_exhibits.forEach((item) => {
-          if (['ttf', 'otf', 'woff'].includes(item.split('.').pop().toLowerCase())) {
-            const option = document.createElement('option')
-            option.value = '../content/' + item
-            option.innerHTML = item
-            $('#fontSelect_' + type).append(option)
-          }
-        })
-      })
-
-      // Then, add the defaults
-      const defaultFonts = ['OpenSans-Light.ttf', 'OpenSans-LightItalic.ttf', 'OpenSans-Regular.ttf', 'OpenSans-Italic.ttf', 'OpenSans-Medium.ttf', 'OpenSans-MediumItalic.ttf', 'OpenSans-SemiBold.ttf', 'OpenSans-SemiBoldItalic.ttf', 'OpenSans-Bold.ttf', 'OpenSans-BoldItalic.ttf', 'OpenSans-ExtraBoldItalic.ttf', 'OpenSans-ExtraBold.ttf']
-
-      types.forEach((type) => {
-        const header = document.createElement('option')
-        header.value = 'Built-in'
-        header.innerHTML = 'Built-in'
-        header.setAttribute('disabled', true)
-        $('#fontSelect_' + type).append(header)
-
-        defaultFonts.forEach((font) => {
-          const option = document.createElement('option')
-          option.value = '../_fonts/' + font
-          option.innerHTML = font
-          $('#fontSelect_' + type).append(option)
-        })
-      })
-    })
-}
-
 // Set helper address for use with constCommon.makeHelperRequest
 constCommon.config.helperAddress = window.location.origin
 
@@ -781,12 +704,9 @@ $('.coloris').change(function () {
   constSetup.updateWorkingDefinition(['style', 'color', $(this).data('property')], value)
   constSetup.previewDefinition(true)
 })
-$('#uploadFontInput').change(onFontUploadChange)
-
-$('.font-select').change(function () {
-  const value = $(this).val().trim()
-  constSetup.updateWorkingDefinition(['style', 'font', $(this).data('property')], value)
-  constSetup.previewDefinition(true)
+document.getElementById('manageFontsButton').addEventListener('click', (event) => {
+  constFileSelect.createFileSelectionModal({ filetypes: ['otf', 'ttf', 'woff', 'woff2'], manage: true })
+    .then(constSetup.refreshAdvancedFontPickers)
 })
 
 // Text size fields
@@ -827,12 +747,12 @@ if (window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').match
 // Set helper address for use with constCommon.makeHelperRequest
 constCommon.config.helperAddress = window.location.origin
 
-populateFontSelects()
 clearDefinitionInput()
 
 constSetup.configure({
   app: 'infostation',
   clearDefinition: clearDefinitionInput,
+  initializeDefinition,
   loadDefinition: editDefinition,
   saveDefinition
 })