+++ control_server/changelog.md
@@ -19,9 +19,11 @@ Constellation 5 introduces a user account system that enables administrators to 
 
 ### General
 - When creating or editing a definition, the preview can be configured for displays of different aspect ratios.
+- Rich previews help you select the perfect font.
 
 ### Media Browser
 - Designate spreadsheet columns as filterable to allow users to sort the media collection.
+- Redesigned lightbox allows media to be bigger.
 
 
 # **Constellation 4**